refactor(github): isolate issue graph seeding
